Changed variable type of employee num

+++ app/src/main/java/com/example/seg_2105_project/Doctor.java
@@ -1,10 +1,10 @@
 package com.example.seg_2105_project;
 
 public class Doctor extends User {
-    long employee_number;
+    int employee_number;
     String[] specialties;
 
-    public Doctor(String firstName, String lastName, String email, String password, int phoneNumber, String address, long employee_number, String[] specialties) {
+    public Doctor(String firstName, String lastName, String email, String password, int phoneNumber, String address, int employee_number, String[] specialties) {
         super(firstName, lastName, email, password, phoneNumber, address);
         this.employee_number = employee_number;
         this.specialties = specialties;
